AgentEconomist/requirements.txt
# LangChain & LangGraph 核心
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-sdk>=0.1.30

# 环境配置
python-dotenv>=1.0.0

# 数据处理
pyyaml>=6.0.1

# 如果使用知识库（Qdrant）
qdrant-client>=1.7.0

# 其他工具
httpx>=0.27.0
